+++ packages/util-crypto/src/base64/pad.spec.ts
@@ -1,0 +1,12 @@
+// Copyright 2017-2020 @polkadot/util-crypto authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { base64Pad } from '.';
+
+describe('base64Pad', (): void => {
+  it('pads a utf-8 string', (): void => {
+    expect(
+      base64Pad('YWJjZA')
+    ).toEqual('YWJjZA==');
+  });
+});