+++ packages/util-crypto/src/base64/pad.ts
@@ -1,0 +1,6 @@
+// Copyright 2017-2020 @polkadot/util-crypto authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export default function base64Trim (value: string): string {
+  return value.padEnd(value.length + (value.length % 4), '=');
+}