+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Contributed:
 Changes:
 
 - Use cross-environment TextDecoder & TextEncoder from the polkadot-js polyfills
+- Add `isBase{32, 58, 64}` to `@polkadot/util-crypto`
+- Add `base64{Pad, Trim}` to add/remove padding characters
 
 
 ## 3.4.1 Aug 31, 2020